+++ changes.md
@@ -3,6 +3,6 @@ Changelog
 
 Dev
 ---
-Fixed a bug in `ast_ctx_fixer`
-Implemented `macropy.case_classes.enum` macro
-
+- Fixed a bug in `ast_ctx_fixer`
+- Implemented `macropy.case_classes.enum` macro
+- Implemented `macropy.quick_lambda.lazy` and `macropy.quick_lambda.interned` macros